api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
openai==1.40.1
pydantic==2.7.1
python-dotenv==1.0.1
chromadb==0.5.3
neo4j==5.28.0
python-multipart==0.0.9
tiktoken==0.7.0
pyyaml>=6.0.0